requirements.txt
Django~=4.0.5
psycopg2-binary==2.8.6
djangorestframework==3.12.4
pytz==2022.1
requests==2.28.1
